# src/main/resources/static/openapi.yaml
openapi: 3.0.1
info:
  title: E-Commerce Platform API
  description: REST API documentation for the E-Commerce platform.
  version: "1.0"
servers:
  - url: http://localhost:8080
    description: Local development server
paths:
...
  /products:
    get:
      tags:
        - Products
      summary: Get all products
      responses:
        '200':
          description: List of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProductResponseDto'
    post:
      tags:
        - Products
      summary: Create a new product (ADMIN only)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductRequestDto'
      responses:
        '201':
          description: Product created
  /products/{id}:
    get:
      tags:
        - Products
      summary: Get a product by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Product details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponseDto'
    put:
      tags:
        - Products
      summary: Update a product (ADMIN only)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductRequestDto'
      responses:
        '200':
          description: Product updated
    delete:
      tags:
        - Products
      summary: Delete a product (ADMIN only)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Product deleted
  /products/category/{id}:
    get:
      tags:
        - Products
      summary: Get products by category ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: List of products in the specified category
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProductResponseDto'
  /products/search/{name}:
    get:
      tags:
        - Products
      summary: Search products by name
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: List of products matching the search query
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProductResponseDto'

  # --- ORDERS ENDPOINTS ---
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ProductRequestDto:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        price:
          type: number
          format: double
        stockQuantity:
          type: integer
          format: int32
        categoryId:
          type: integer
          format: int64
        imageUrl:
          type: string
    ProductResponseDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        description:
          type: string
        price:
          type: number
          format: double
        stockQuantity:
          type: integer
          format: int32
        categoryName:
          type: string
        imageUrl:
          type: string
...